refactor(ios-keyboard): split top-row accessory layout
- Move top-row accessory positioning out of KeyboardLayoutGeometry into KeyboardTopRowAccessoryLayout
- Keep KeyboardLayoutGeometry focused on row width calculations
- Convert single-key accessories to a descriptor-driven layout path for future keyboard controls
- Keep Vibes, logo, and cancel positioning explicit because they have custom layout behavior
- Update keyboard layout docs for the new ownership split
